services/api/pyproject.toml
[project]
name = "ai-data-factory-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "pydantic==2.6.4",
  "sqlalchemy==2.0.29",
  "alembic==1.13.1",
  "psycopg2-binary==2.9.9",
  "python-multipart==0.0.9",
  "minio==7.2.5",
  "opensearch-py==2.4.2",
  "httpx==0.27.0",
  "langfuse==2.0.15",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "promptfoo==1.12.0",
  "pytest==8.2.0",
  "ruff==0.4.9"
]
